fix(test): narrow parent quote probe explicitly
- Cast the bounded probe amount to the quoter's uint128 ABI width
- Keep public test constants wide while narrowing only at the call boundary

+++ test/liquidity/fork/RobinhoodNestedFlashArbitrageFork.t.sol
@@ -150,7 +150,7 @@ contract RobinhoodNestedFlashArbitrageForkTest is RobinhoodNestedBasketsForkBase
         );
         (amountOut,) = quoter.quoteExactInputSingle(
             IV4Quoter.QuoteExactSingleParams({
-                poolKey: pool, zeroForOne: zeroForOne, exactAmount: PURCHASE_PROBE, hookData: ""
+                poolKey: pool, zeroForOne: zeroForOne, exactAmount: uint128(PURCHASE_PROBE), hookData: ""
             })
         );
     }
